+++ app/javascript/helpers/objectKeysToSnakeCase.js
@@ -1,7 +1,8 @@
 export const objectKeysToSnakeCase = (source) =>
   Object.keys(source)
-    .reduce((destination, key) => {
+    .reduce((acc, key) => {
+      let valueIsObject = typeof source[key] === 'object' && source[key] !== null;
       let transformedKey = key.replace(/([A-Z])/g, "_$1").toLowerCase();
-      destination[transformedKey] = source[key];
-      return destination;
+      acc[transformedKey] = valueIsObject ? objectKeysToSnakeCase(source[key]) : source[key];
+      return acc;
     }, {});